make lookup table threshold a floating point value
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@32871 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/algo/blast/api/blast_options_local_priv.hpp
@@ -71,8 +71,8 @@ public:
     void SetProgram(EProgram p);
 
     /******************* Lookup table options ***********************/
-    int GetWordThreshold() const;
-    void SetWordThreshold(int w);
+    double GetWordThreshold() const;
+    void SetWordThreshold(double w);
 
     ELookupTableType GetLookupTableType() const;
     void SetLookupTableType(ELookupTableType type);
@@ -389,14 +389,14 @@ CBlastOptionsLocal::SetMatrixName(const char* matrix)
     m_ScoringOpts->matrix = strdup(matrix);
 }
 
-inline int
+inline double
 CBlastOptionsLocal::GetWordThreshold() const
 {
     return m_LutOpts->threshold;
 }
 
 inline void
-CBlastOptionsLocal::SetWordThreshold(int w)
+CBlastOptionsLocal::SetWordThreshold(double w)
 {
     m_LutOpts->threshold = w;
 }
