+++ src/algo/blast/api/demo/blast2seq.cpp
@@ -260,7 +260,7 @@ CBlast2seqApplication::ProcessCommandLineArgs()
         opt.SetMBTemplateType(args["templtype"].AsInteger());
     }
     if (args["thresh"].AsInteger()) {
-        opt.SetWordThreshold(args["thresh"].AsInteger());
+        opt.SetWordThreshold(args["thresh"].AsDouble());
     }
     if (args["window"].AsInteger()) {
         opt.SetWindowSize(args["window"].AsInteger());
